+++ dsound/AddressLookupTable.h
@@ -43,45 +43,45 @@ public:
 	template <typename T>
 	struct AddressCacheIndex {};
 	template <>
-	struct AddressCacheIndex<m_IKsPropertySet> { static constexpr UINT CacheIndex = 0; };
+	struct AddressCacheIndex<m_IKsPropertySet> { static constexpr size_t CacheIndex = 0; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSound8> { static constexpr UINT CacheIndex = 1; };
+	struct AddressCacheIndex<m_IDirectSound8> { static constexpr size_t CacheIndex = 1; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSound3DBuffer8> { static constexpr UINT CacheIndex = 2; };
+	struct AddressCacheIndex<m_IDirectSound3DBuffer8> { static constexpr size_t CacheIndex = 2; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSound3DListener8> { static constexpr UINT CacheIndex = 3; };
+	struct AddressCacheIndex<m_IDirectSound3DListener8> { static constexpr size_t CacheIndex = 3; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundBuffer8> { static constexpr UINT CacheIndex = 4; };
+	struct AddressCacheIndex<m_IDirectSoundBuffer8> { static constexpr size_t CacheIndex = 4; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundCapture8> { static constexpr UINT CacheIndex = 5; };
+	struct AddressCacheIndex<m_IDirectSoundCapture8> { static constexpr size_t CacheIndex = 5; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundCaptureBuffer8> { static constexpr UINT CacheIndex = 6; };
+	struct AddressCacheIndex<m_IDirectSoundCaptureBuffer8> { static constexpr size_t CacheIndex = 6; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundCaptureFXAec8> { static constexpr UINT CacheIndex = 7; };
+	struct AddressCacheIndex<m_IDirectSoundCaptureFXAec8> { static constexpr size_t CacheIndex = 7; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundCaptureFXNoiseSuppress8> { static constexpr UINT CacheIndex = 8; };
+	struct AddressCacheIndex<m_IDirectSoundCaptureFXNoiseSuppress8> { static constexpr size_t CacheIndex = 8; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFullDuplex8> { static constexpr UINT CacheIndex = 9; };
+	struct AddressCacheIndex<m_IDirectSoundFullDuplex8> { static constexpr size_t CacheIndex = 9; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXChorus8> { static constexpr UINT CacheIndex = 10; };
+	struct AddressCacheIndex<m_IDirectSoundFXChorus8> { static constexpr size_t CacheIndex = 10; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXCompressor8> { static constexpr UINT CacheIndex = 11; };
+	struct AddressCacheIndex<m_IDirectSoundFXCompressor8> { static constexpr size_t CacheIndex = 11; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXDistortion8> { static constexpr UINT CacheIndex = 12; };
+	struct AddressCacheIndex<m_IDirectSoundFXDistortion8> { static constexpr size_t CacheIndex = 12; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXEcho8> { static constexpr UINT CacheIndex = 13; };
+	struct AddressCacheIndex<m_IDirectSoundFXEcho8> { static constexpr size_t CacheIndex = 13; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXFlanger8> { static constexpr UINT CacheIndex = 14; };
+	struct AddressCacheIndex<m_IDirectSoundFXFlanger8> { static constexpr size_t CacheIndex = 14; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXGargle8> { static constexpr UINT CacheIndex = 15; };
+	struct AddressCacheIndex<m_IDirectSoundFXGargle8> { static constexpr size_t CacheIndex = 15; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXI3DL2Reverb8> { static constexpr UINT CacheIndex = 16; };
+	struct AddressCacheIndex<m_IDirectSoundFXI3DL2Reverb8> { static constexpr size_t CacheIndex = 16; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXParamEq8> { static constexpr UINT CacheIndex = 17; };
+	struct AddressCacheIndex<m_IDirectSoundFXParamEq8> { static constexpr size_t CacheIndex = 17; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundFXWavesReverb8> { static constexpr UINT CacheIndex = 18; };
+	struct AddressCacheIndex<m_IDirectSoundFXWavesReverb8> { static constexpr size_t CacheIndex = 18; };
 	template <>
-	struct AddressCacheIndex<m_IDirectSoundNotify8> { static constexpr UINT CacheIndex = 19; };
+	struct AddressCacheIndex<m_IDirectSoundNotify8> { static constexpr size_t CacheIndex = 19; };
 
 	template <typename T>
 	T *FindAddress(void *Proxy)
